Memory management, freeing

+++ app/objects/obj_cube.h
@@ -6,7 +6,8 @@
 #define DEEPCARS_OBJ_CUBE_H
 
 #include "../../lib/objectInclude.h"
+#include "../winDefaults.h"
 
-drawableObject_t* create_cube(vec3f_t pos, double size, material_t* mat);
+drawableObject_t* create_cube(vec3f_t pos, float size, material_t* mat);
 
 #endif //DEEPCARS_OBJ_CUBE_H
